updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -32,8 +32,10 @@ with app.app_context():
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user = db.session.query(User).filter_by(email='nfinnemo@uncc.edu').one()
-    return render_template('index.html', user = a_user)
+    if session.get('user'):
+        return render_template('index.html', user = session['user'])
+    return render_template("index.html")
+   
 
 @app.route("/notes")
 def get_notes():
@@ -54,49 +56,52 @@ def get_note(note_id):
 
 @app.route("/notes/new", methods=['GET', 'POST'])
 def new_note():
-    
-    if request.method == 'POST':
-        title = request.form['title']
-        text = request.form['noteText']
-        from datetime import date
-        today = date.today()
-        new_record = Note(title, text, today)
-        db.session.add(new_record)
-        db.session.commit()
-        return redirect(url_for('get_notes'))
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
+            text = request.form['noteText']
+            from datetime import date
+            today = date.today()
+            new_record = Note(title, text, today, session['user_id'])
+            db.session.add(new_record)
+            db.session.commit()
+            return redirect(url_for('get_notes'))
+        else:
+            return render_template('new.html', user=session['user'])
     else:
-        a_user = db.session.query(User).filter_by(email='nfinnemo@uncc.edu').one()
-        return render_template('new.html', user=a_user)
+        return redirect(url_for('login'))
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
+    if session.get('user'):
     #check method used for request
-    if request.method == 'POST':
-        title=request.form['title']
-        text = request.form['noteText']
-        note=db.session.query(Note).filter_by(id=note_id).one()
-        note.title = title
-        note.text = text
-        db.session.add(note)
-        db.session.commit()
-        return redirect(url_for('get_notes'))
+        if request.method == 'POST':
+            title=request.form['title']
+            text = request.form['noteText']
+            note=db.session.query(Note).filter_by(id=note_id).one()
+            note.title = title
+            note.text = text
+            db.session.add(note)
+            db.session.commit()
+            return redirect(url_for('get_notes'))
+        else:
+            #GET request - show new note form to edit note
+            # retrieve user from database
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+            return render_template('new.html', note=my_note, user=session['user'])
     else:
-        #GET request - show new note form to edit note
-        # retrieve user from database
-        a_user = db.session.query(User).filter_by(email='nfinnemo@uncc.edu').one()
-
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note=my_note, user=a_user)
+        return redirect(url_for('login'))
 
 @app.route('/notes/delete/<note_id>', methods=['POST'])
 def delete_note(note_id):
+    if session.get('user'):
     #get note from database
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
-
-    return redirect(url_for('get_notes'))
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
+        db.session.commit()
+        return redirect(url_for('get_notes'))
+    else:
+        return redirect(url_for('login'))
 
 @app.route('/register', methods=['POST', 'GET'])
 def register():
